--- src/components/profile/profile-avatar.tsx
@@ -1,14 +1,0 @@
-import { FaUser } from 'react-icons/fa'
-
-import { Avatar, AvatarFallback, AvatarImage } from '@/components/ui/avatar'
-
-export const ProfileAvatar = () => {
-  return (
-    <Avatar>
-      <AvatarImage />
-      <AvatarFallback className="text-primary">
-        <FaUser size="24" />
-      </AvatarFallback>
-    </Avatar>
-  )
-}
